Add the South African Identity Federation
We are currently producing a local JSON feed for DiscoJuice to allow
IdPs that have opted not to participate in eduGAIN to make use of this.
However, the local feed format is limited and doesn't allow logos, etc.
So adding the federation here explicitly would allow for a more
consistant look and feel.

Metadata URL as documented at https://safire.ac.za/technical/metadata/